clothe/requirements.txt
appdirs==1.4.3
beautifulsoup4==4.8.2
bs4==0.0.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
Django==1.11.20
fake-useragent==0.1.11
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
grpcio==1.26.0
idna==2.8
lxml==4.4.2
mccabe==0.6.1
parse==1.14.0
Pillow==6.2.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyee==6.0.0
pyppeteer==0.0.25
pyquery==1.4.1
pytz==2019.3
requests==2.22.0
requests-html==0.10.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
tqdm==4.41.0
urllib3==1.25.7
w3lib==1.21.0
websockets==8.1
wincertstore==0.2
